Refactor to simplify dependency verification code
- Only the linker trace needs to handle the case where XBS is setting
LD_TRACE_FILE and needs to have all the trace info combined, so put the code
for that into LdTaskAction instead of sharing it for all dependency
verification.
- With that change, there is not much left for the TaskDependencyVerification
Adapter protocol to handle, so we can just remove that. This will also make
it easier to handle dependency in the existing ClangCompileTaskAction code
for module-enabled compilations.
- Remove TaskExecutionContext, since that is no longer needed.
- Revert the change to make TaskAction's spawn function static.
- This leaves the TaskDependencyVerification struct with only some static
functions. This seems like a good place to share code for checking the various
different kinds of dependencies, so I am leaving it as-is for now.