Align this branch with the pr-swift-mod-deps branch
This pulls in some changes from the pr-swift-mod-deps branch (in
anticipation of that branch getting merged before this one).

- The MODULE_DEPENDENCIES setting has already been added, so use that
instead of the DEPENDENCIES setting that was in this branch. There is
more to do also read libraries from OTHER_LDFLAGS and the link phases.

- Similarly, use VALIDATE_MODULE_DEPENDENCIES instead of the
DEPENDENCIES_VERIFICATION setting from this branch.

- Add a similar VALIDATE_LINK_DEPENDENCIES setting.

- Replace TaskDependencySettings with the ModuleDependenciesContext struct
from the other branch and add a similar LinkDependenciesContext struct.

- Rename the default LD_TRACE_FILE to use a ".ld_trace.json" suffix.
Since it is located in the object file directory, it seems good to more
clearly distinguish it from the Clang trace files ending with ".trace.json".